page_insights_by_day method
response value of an insights request can be a hash instead of number
value. This hash was a nested hash sometimes. For that reason, we
need a method returns the whole array, and the let a rails app can
use deep_merge on it. But we can also require activesupport and use
deep_merge from fb-core. I'm not sure what is better.